requirements.txt
Django==3.1.2
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.1
gunicorn==21.2.0
